refactor(app): split event handler navigation and unit panel modules (#4224)
Extract dialog/navigation routing and shared units-panel sheet host from
app_event_handler_unit_panels.dart into focused sibling libraries so the
unit-panels module is ≤250 physical lines per wave-11 Slice C.
